+++ CHANGELOG.md
@@ -1,2 +1,7 @@
 ### 1.0.2
 Updated `peerDependencies` to allow React 15 or 16. Either is fine.
+
+### 1.1.0
+Enabled `NODE_ENV=TEST` mode, which removes the requirment that
+components be run with context (i.e., no need for a
+`StyletronProvider` or `ThemeProvider`.)